Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
google-generativeai==0.3.2
Pillow==10.1.0
aiofiles==23.2.1
